OC-9772 RT does not logout
