+++ web/src/main/java/org/akaza/openclinica/config/AppConfig.java
@@ -79,6 +79,7 @@ public class AppConfig extends KeycloakWebSecurityConfigurerAdapter {
                 .antMatchers("/css/**","/includes/**","/images/**", "/fonts/**",
                         "/js/**",
                         "/callback", "/sso/login",
+                        "/pages/login",
                         "/pages/customer-service/**",
                         "/pages/ocLogin",
                         "/pages/resetOCAppTimeout",
